Renames spatial_vectors.h --> spatial_algebra.h
