+++ drake/multibody/multibody_tree/math/spatial_algebra.h
